chore: cleanup unused files and migrate to ts
