--- src/shared/store/authStore.js
@@ -1,30 +1,0 @@
-// -----------------------------------------------------------------------------
-// Zustand-хранилище профиля + переключение проекта
-// -----------------------------------------------------------------------------
-import { create } from 'zustand';
-
-/** Zustand-store профиля текущего пользователя */
-export const useAuthStore = create((set) => ({
-    /** undefined → ещё грузится; null → гость; object → авторизован            */
-    profile: undefined,
-
-    /** Сохраняем весь профиль, полученный с сервера -------------------------*/
-    setProfile: (profile) => set({ profile }),
-
-    /** Очистка при выходе ---------------------------------------------------*/
-    clearProfile: () => set({ profile: null }),
-
-    /** ----------------------------------------------------------------------
-     * CHANGE: обновляем только project_id, сохраняя остальные поля профиля
-     * благодаря этому селекторы (useProjectId) и React-Query реактивно
-     * перерисуются без полной перезагрузки страницы
-     * ---------------------------------------------------------------------*/
-    setProjectId: (project_id) =>
-        set((s) => (s.profile
-            ? { profile: { ...s.profile, project_id } }
-            : {})),
-}));
-
-/** Селектор project_id (null, если пользователь гость либо проект не назначен) */
-export const useProjectId = () =>
-    useAuthStore((s) => s.profile?.project_id ?? null);
